+++ src/final_login/routers/ml.py
@@ -1,8 +1,8 @@
-from fastapi import FastAPI, HTTPException, APIRouter
+from fastapi import HTTPException, APIRouter
 from pydantic import BaseModel
 from motor.motor_asyncio import AsyncIOMotorClient
 from bson import ObjectId
-from typing import List,Optional
+from typing import List
 from dotenv import load_dotenv
 import os
